Remove test cases for protocols that don't support event streams
