+++ codegen-server/src/test/kotlin/software/amazon/smithy/rust/codegen/server/smithy/protocols/eventstream/ServerEventStreamUnmarshallerGeneratorTest.kt
@@ -5,10 +5,7 @@
 
 package software.amazon.smithy.rust.codegen.server.smithy.protocols.eventstream
 
-import org.junit.jupiter.api.extension.ExtensionContext
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
 import software.amazon.smithy.codegen.core.Symbol
 import software.amazon.smithy.model.shapes.StructureShape
@@ -19,29 +16,22 @@ import software.amazon.smithy.rust.codegen.core.smithy.generators.BuilderGenerat
 import software.amazon.smithy.rust.codegen.core.smithy.generators.implBlock
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.Protocol
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.parse.EventStreamUnmarshallerGenerator
-import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestModels
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestTools
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestVariety
 import software.amazon.smithy.rust.codegen.core.testutil.TestEventStreamProject
 import software.amazon.smithy.rust.codegen.server.smithy.ServerCodegenContext
 import software.amazon.smithy.rust.codegen.server.smithy.generators.serverBuilderSymbol
-import java.util.stream.Stream
-
-class UnmarshallTestCasesProvider : ArgumentsProvider {
-    override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> =
-        EventStreamTestModels.TEST_CASES.flatMap { testCase ->
-            listOf(
-                // TODO(https://github.com/awslabs/smithy-rs/issues/1442): Enable tests for `publicConstrainedTypes = false`
-                // TestCase(testCase, publicConstrainedTypes = false),
-                TestCase(testCase, publicConstrainedTypes = true),
-            )
-        }.map { Arguments.of(it) }.stream()
-}
 
 class ServerEventStreamUnmarshallerGeneratorTest {
     @ParameterizedTest
-    @ArgumentsSource(UnmarshallTestCasesProvider::class)
+    @ArgumentsSource(TestCasesProvider::class)
     fun test(testCase: TestCase) {
+        // TODO(https://github.com/awslabs/smithy-rs/issues/1442): Enable tests for `publicConstrainedTypes = false`
+        // by deleting this if/return
+        if (!testCase.publicConstrainedTypes) {
+            return
+        }
+
         EventStreamTestTools.runTestCase(
             testCase.eventStreamTestCase,
             object : ServerEventStreamBaseRequirements() {
